requirements.txt
jsonschema
openai
pytest
python-docx
streamlit
scikit-learn
PyPDF2
